Extract default function test into a separate class
Since new string functions are added into an extended function registry
it requires different runtime created to test those new functions. On
the other hand, helper functions will definitely come handy when testing
the string function registry. This is the main reason of this separation
though the class names may by far not the best.
