+++ jmespath-core/src/test/java/io/burt/jmespath/jcf/JcfTest.java
@@ -1,25 +1,32 @@
 package io.burt.jmespath.jcf;
 
-import java.util.List;
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 
 import org.junit.Test;
+import org.junit.experimental.runners.Enclosed;
+import org.junit.runner.RunWith;
 
-import io.burt.jmespath.JmesPathRuntimeTest;
-import io.burt.jmespath.RuntimeConfiguration;
 import io.burt.jmespath.Adapter;
+import io.burt.jmespath.JmesPathRuntimeWithDefaultConfigurationTest;
+import io.burt.jmespath.RuntimeConfiguration;
 
-import static org.junit.Assert.assertThat;
-import static org.hamcrest.Matchers.is;
+@RunWith(Enclosed.class)
+public class JcfTest {
+  public static class DefaultConfiguration extends JmesPathRuntimeWithDefaultConfigurationTest<Object> {
+    @Override
+    protected Adapter<Object> createRuntime(RuntimeConfiguration configuration) { return new JcfRuntime(configuration); }
 
-public class JcfTest extends JmesPathRuntimeTest<Object> {
-  @Override
-  protected Adapter<Object> createRuntime(RuntimeConfiguration configuration) { return new JcfRuntime(configuration); }
+    @Test
+    public void toListReturnsAListWhenGivenAnotherTypeOfCollection() {
+      List<Object> list = runtime().toList(Collections.singleton(parse("1")));
+      assertThat(list, is(Arrays.asList(parse("1"))));
+    }
+  }
 
-  @Test
-  public void toListReturnsAListWhenGivenAnotherTypeOfCollection() {
-    List<Object> list = runtime().toList(Collections.singleton(parse("1")));
-    assertThat(list, is(Arrays.asList(parse("1"))));
   }
 }
