+++ jmespath-gson/src/test/java/io/burt/jmespath/gson/GsonTest.java
@@ -1,11 +1,18 @@
 package io.burt.jmespath.gson;
 
+import org.junit.experimental.runners.Enclosed;
+import org.junit.runner.RunWith;
+
 import com.google.gson.JsonElement;
+
 import io.burt.jmespath.Adapter;
-import io.burt.jmespath.JmesPathRuntimeTest;
+import io.burt.jmespath.JmesPathRuntimeWithDefaultConfigurationTest;
 import io.burt.jmespath.RuntimeConfiguration;
 
-public class GsonTest extends JmesPathRuntimeTest<JsonElement> {
-  @Override
-  protected Adapter<JsonElement> createRuntime(RuntimeConfiguration configuration) { return new GsonRuntime(configuration); }
+@RunWith(Enclosed.class)
+public class GsonTest {
+  public static class DefaultConfiguration extends JmesPathRuntimeWithDefaultConfigurationTest<JsonElement> {
+    @Override
+    protected Adapter<JsonElement> createRuntime(RuntimeConfiguration configuration) { return new GsonRuntime(configuration); }
+  }
 }
