+++ jmespath-jackson/src/test/java/io/burt/jmespath/jackson/JacksonTest.java
@@ -1,12 +1,18 @@
 package io.burt.jmespath.jackson;
 
+import org.junit.experimental.runners.Enclosed;
+import org.junit.runner.RunWith;
+
 import com.fasterxml.jackson.databind.JsonNode;
 
-import io.burt.jmespath.JmesPathRuntimeTest;
+import io.burt.jmespath.JmesPathRuntimeWithDefaultConfigurationTest;
 import io.burt.jmespath.RuntimeConfiguration;
 import io.burt.jmespath.Adapter;
 
-public class JacksonTest extends JmesPathRuntimeTest<JsonNode> {
-  @Override
-  protected Adapter<JsonNode> createRuntime(RuntimeConfiguration configuration) { return new JacksonRuntime(configuration); }
+@RunWith(Enclosed.class)
+public class JacksonTest {
+  public static class DefaultConfiguration extends JmesPathRuntimeWithDefaultConfigurationTest<JsonNode> {
+    @Override
+    protected Adapter<JsonNode> createRuntime(RuntimeConfiguration configuration) { return new JacksonRuntime(configuration); }
+  }
 }
